refactor: 불변 객체 vo 패키지로 이동
* 불변 객체 record로 변경
