+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.vo.Lotto;
+
 public interface LottoGenerator {
 
     Lotto generate();
