+++ src/main/java/lotto/domain/vo/Lottos.java
@@ -1,17 +1,16 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Lottos {
-
-    private final List<Lotto> lottos;
+public record Lottos(List<Lotto> lottos) {
 
     public Lottos(List<Lotto> lottos) {
         this.lottos = List.copyOf(lottos);
     }
 
-    public List<Lotto> getLottos() {
+    @Override
+    public List<Lotto> lottos() {
         return new ArrayList<>(lottos);
     }
 }
