+++ src/main/java/lotto/domain/vo/Money.java
@@ -1,14 +1,11 @@
-package lotto.domain;
+package lotto.domain.vo;
 
-public class Money {
+public record Money(int money) {
 
     private static final int UNIT = 1000;
 
-    private final int money;
-
-    public Money(int money) {
+    public Money {
         validateMoney(money);
-        this.money = money;
     }
 
     private static void validateMoney(int money) {
@@ -22,9 +19,5 @@ public class Money {
 
     public int calculateQuantity(int unitPrice) {
         return money / unitPrice;
-    }
-
-    public int getMoney() {
-        return money;
     }
 }
