+++ src/test/java/lotto/domain/WinningLottoTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import lotto.domain.vo.Lotto;
+import lotto.domain.vo.Rank;
 import org.junit.jupiter.api.Test;
 
 class WinningLottoTest {
